pyproject.toml
[tool.poetry]
name = "basefastapimvc"
version = "0.1.0"
description = "FastAPI MVC application with service layer"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.116.1"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.23"
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.3.0"
python-dotenv = "^1.0.0"
alembic = "^1.12.1"
werkzeug = "^3.0.0"
celery = "^5.3.4"
redis = "^5.0.1"
apscheduler = "^3.10.4"
loguru = "^0.7.2"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
asyncpg = "^0.30.0"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest-cov = "^4.1.0"
pytest-asyncio = "<1.0.0"
aiosqlite = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
